* fix para que desde la operación alta manual de pagos, no de un notice al querer eliminar la columna grado * se agrega la lógica para que a los tutores les aparezca una notificación al inicio

+++ php/personas/vincular_allegados/ci_vincular_allegados.php
@@ -80,8 +80,10 @@ class ci_vincular_allegados extends ci_abm_personas
             $this->s__datos_filtro['vincular_allegados'] = true;
         }
         if (isset($this->s__datos_filtro)) {
-            if (!$this->s__datos_filtro['nivel']) {
-                $cuadro->eliminar_columnas(array('grado'));
+            if ($cuadro->existe_columna('grado')) {
+                if (!$this->s__datos_filtro['nivel']) {
+                    $cuadro->eliminar_columnas(array('grado'));
+                }
             }
         }
         parent::conf__cuadro($cuadro);
